edit price storage in db

+++ db/migrate/20190425221154_create_finds.rb
@@ -4,7 +4,7 @@ class CreateFinds < ActiveRecord::Migration[5.2]
       t.integer :user_id
       t.integer :store_id
       t.string :photo
-      t.integer :price
+      t.money :price
       t.string :brand
       t.text :description
